[metasrv] refactor: move raft state into metasrv/raft/state/
