+++ metasrv/src/meta_service/raftmeta.rs
@@ -52,7 +52,6 @@ use common_tracing::tracing;
 use common_tracing::tracing::Instrument;
 
 use crate::configs;
-use crate::meta_service::raft_state::RaftState;
 use crate::meta_service::Cmd;
 use crate::meta_service::LogEntry;
 use crate::meta_service::MetaServiceClient;
@@ -62,6 +61,7 @@ use crate::meta_service::Network;
 use crate::meta_service::RetryableError;
 use crate::meta_service::ShutdownError;
 use crate::raft::log::RaftLog;
+use crate::raft::state::RaftState;
 use crate::raft::state_machine::AppliedState;
 use crate::raft::state_machine::Node;
 use crate::raft::state_machine::SerializableSnapshot;
