+++ metasrv/src/sled_store/sled_key_space.rs
@@ -28,8 +28,8 @@ use sled::IVec;
 use crate::meta_service::LogEntry;
 use crate::meta_service::LogIndex;
 use crate::meta_service::NodeId;
-use crate::meta_service::RaftStateKey;
-use crate::meta_service::RaftStateValue;
+use crate::raft::state::RaftStateKey;
+use crate::raft::state::RaftStateValue;
 use crate::raft::state_machine::Node;
 use crate::raft::state_machine::StateMachineMetaKey;
 use crate::raft::state_machine::StateMachineMetaValue;
